refactor and delete empty files
