--- lib/core/constants/constants.dart
@@ -1,15 +1,0 @@
-abstract class Constants {
-  static const String token = 'token';
-  static const String isTokenSaved = 'isTokenSaved';
-
-  //register
-  static const String firstName = 'firstName';
-  static const String lastName = 'lastName';
-  static const String email = 'email';
-  static const String password = 'password';
-  static const String rePassword = 'rePassword';
-  static const String phone = 'phone';
-  static const String gender = 'gender';
-
-  static const String noInternet = 'No Internet Connection';
-}
