+++ lib/core/utils/app_constants.dart
@@ -1,3 +1,15 @@
 abstract class AppConstants {
+  static const String token = 'token';
+  static const String isTokenSaved = 'isTokenSaved';
 
+  //register
+  static const String firstName = 'firstName';
+  static const String lastName = 'lastName';
+  static const String email = 'email';
+  static const String password = 'password';
+  static const String rePassword = 'rePassword';
+  static const String phone = 'phone';
+  static const String gender = 'gender';
+
+  static const String noInternet = 'No Internet Connection';
 }
