+++ lib/core/utils/json_serializable_constants.dart
